types/rect.rs: Implement Rect & SplitRect
This commit adds Rect and SplitRect to represent area.
Rects are straightforward and represent the area with a pair of
coordinate.
SplitRect wrap a vector of Rects, and can be use to 'mask' part of a
rects. The 'bound()' methode return a new Rect which is a bounding box
around every Rects in the SplitRect.

CHANGELOG:
- Adds Rect, a pair of coordinate representing a regions of space.
- Adds SplitRect, a wrapper around a vector of Rectangle.